Adding operation class to a new file
